+++ DBCD.IO/Attributes/IndexAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace DBFileReaderLib.Attributes
+namespace DBCD.IO.Attributes
 {
     public class IndexAttribute : Attribute
     {
