+++ DBCD.IO/Extensions.cs
@@ -6,7 +6,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Text;
 
-namespace DBFileReaderLib
+namespace DBCD.IO
 {
     static class Extensions
     {
