+++ DBCD.IO/Readers/WDC2Reader.cs
@@ -1,4 +1,4 @@
-﻿using DBFileReaderLib.Common;
+﻿using DBCD.IO.Common;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Text;
 
-namespace DBFileReaderLib.Readers
+namespace DBCD.IO.Readers
 {
     class WDC2Row : IDBRow
     {
